requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
kafka-python
python-dotenv